Removed 1.7-1.10 support; fixed typo; redid pom.xml
Maven doesn't allow for multiple MC versions to be specified in the
dependencies, which explains the original author's decision to use the
(rather confusing) modules system. This version has been simplified, but
at the cost of dropping support for all MC versions other than 1.12 and
1.12.2. The POM file was switched to build JAR outputs and include
plugin.yml in them. This will likely be the last update to this
repository.